#Purchase Invoice Changes

+++ Client/Components/EditableInvoiceItems.razor.cs
@@ -15,7 +15,7 @@ public partial class EditableInvoiceItems
     [Parameter] public EventCallback<decimal> TotalAmountChanged { get; set; }
     [Parameter] public EventCallback<decimal> TotalTaxChanged { get; set; }
     [Parameter] public EventCallback<decimal> DiscountChanged { get; set; }
-    [Parameter] public EventCallback<double> DiscountPercentageChanged { get; set; }
+    [Parameter] public EventCallback<decimal> DiscountPercentageChanged { get; set; }
     [Parameter] public EventCallback<DiscountType> DiscountTypeChanged { get; set; }
     [Parameter] public EventCallback<decimal> SubTotalChanged { get; set; }
     private RadzenDataGrid<InvoiceItemDTO> itemsGrid = null!;
@@ -23,7 +23,7 @@ public partial class EditableInvoiceItems
     public decimal SubTotalAmount => Items?.Where(i => !i.IsDeleted).Sum(i => i.Total) ?? 0;
     public decimal TaxAmount => Items?.Where(i => !i.IsDeleted).Sum(i => i.TaxAmount) ?? 0;
     public decimal TotalAmount => Items?.Where(i => !i.IsDeleted).Sum(i => i.Total) ?? 0;
-    public double DiscountPercentage { get; set; } = 0;
+    public decimal DiscountPercentage { get; set; } = 0;
     public decimal DiscountAmt { get; set; } = 0;
     public decimal GrandTotalAmount { get; set; } = 0;
 
@@ -31,7 +31,7 @@ public partial class EditableInvoiceItems
     {
         if (Invoice != null)
         {
-            DiscountPercentage = (double)Invoice.DiscountPercent;
+            DiscountPercentage = Invoice.DiscountPercent;
             DiscountAmt = Invoice.DiscountAmount;
             GrandTotalAmount = Invoice.TotalAmount;
         }
