+++ Objects/DTOs/JobBulkSaveDTO.cs
@@ -26,7 +26,7 @@ public class JobBulkSaveDTO
 
     public decimal TotalAmount { get; set; }
     public decimal DiscountAmount { get; set; }
-    public double DiscountPercentage { get; set; }
+    public decimal DiscountPercentage { get; set; }
     public decimal CGSTAmount { get; set; } = 0;
     public decimal SGSTAmount { get; set; } = 0;
     public decimal IGSTAmount { get; set; } = 0;
